bug fix: memory unsafety due to incorrect mocks
the VGA constructor is unsafe because it accepts a pointer to an
arbitraty memory location. the compiler does not, and cannot check
whether this location is large enough for what you want to write.

because this is not checked by the compiler, we had skipped the `*2`
which calculates the bytesize of the location and as a result were
pointing to a location half the size of what we needed.